Backend2/requirements.txt
marshmallow>=3
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
uwsgi==2.0.20
psycopg2-binary
flask-cors
minio
pydantic
Pillow
Unidecode
redis
pandas
openpyxl
zipp
pytz
sentry-sdk[flask]
xlrd
# celery[redis]>=5.0.0
elastic-apm[flask]
netifaces
aiosmtplib
apscheduler
flask[async]
requests
firebase-admin